Rename caches to convey their job better
