+++ nanoc/lib/nanoc/base/repos.rb
@@ -6,9 +6,9 @@ Nanoc::DataSource = Nanoc::Core::DataSource
 require_relative 'repos/store'
 
 require_relative 'repos/checksum_store'
-require_relative 'repos/binary_content_cache'
+require_relative 'repos/binary_compiled_content_cache'
+require_relative 'repos/textual_compiled_content_cache'
 require_relative 'repos/compiled_content_cache'
-require_relative 'repos/composite_cache'
 require_relative 'repos/config_loader'
 require_relative 'repos/dependency_store'
 require_relative 'repos/item_rep_repo'
