+++ nanoc/nanoc.manifest
@@ -22,10 +22,10 @@ lib/nanoc/base/feature.rb
 lib/nanoc/base/repos.rb
 lib/nanoc/base/repos/action_sequence_store.rb
 lib/nanoc/base/repos/aggregate_data_source.rb
-lib/nanoc/base/repos/binary_content_cache.rb
+lib/nanoc/base/repos/binary_compiled_content_cache.rb
 lib/nanoc/base/repos/checksum_store.rb
 lib/nanoc/base/repos/compiled_content_cache.rb
-lib/nanoc/base/repos/composite_cache.rb
+lib/nanoc/base/repos/textual_compiled_content_cache.rb
 lib/nanoc/base/repos/config_loader.rb
 lib/nanoc/base/repos/dependency_store.rb
 lib/nanoc/base/repos/in_mem_data_source.rb
