+++ nanoc/spec/nanoc/base/repos/textual_compiled_content_cache_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-describe Nanoc::Int::CompiledContentCache do
+describe Nanoc::Int::TextualCompiledContentCache do
   let(:cache) { described_class.new(config: config) }
 
   let(:items) { [item] }
